+++ hoot-core-test/src/test/cpp/hoot/core/visitors/RemoveTagsVisitorTest.cpp
@@ -34,7 +34,7 @@
 // hoot
 #include <hoot/core/elements/OsmMap.h>
 #include <hoot/core/TestUtils.h>
-#include <hoot/core/criterion/ElementTypeCriterion.h>
+#include <hoot/core/criterion/NodeCriterion.h>
 #include <hoot/core/io/OsmMapReaderFactory.h>
 #include <hoot/core/io/OsmMapWriterFactory.h>
 #include <hoot/core/visitors/RemoveTagsVisitor.h>
