+++ hoot-core-test/src/test/cpp/hoot/core/visitors/TransliterateNameVisitorTest.cpp
@@ -29,7 +29,6 @@
 #include <hoot/core/elements/OsmMap.h>
 #include <hoot/core/TestUtils.h>
 #include <hoot/core/criterion/PoiCriterion.h>
-#include <hoot/core/criterion/HighwayCriterion.h>
 #include <hoot/core/io/OsmJsonWriter.h>
 #include <hoot/core/io/OsmMapReaderFactory.h>
 #include <hoot/core/io/OsmXmlWriter.h>
