+++ hoot-core/src/main/cpp/hoot/core/algorithms/changeset/ChangesetReplacementElementIdSynchronizer.h
@@ -30,7 +30,6 @@
 
 // Hoot
 #include <hoot/core/elements/ElementIdSynchronizer.h>
-#include <hoot/core/criterion/WayNodeCriterion.h>
 
 namespace hoot
 {
