+++ hoot-core/src/main/cpp/hoot/core/conflate/SuperfluousConflateOpRemover.cpp
@@ -28,7 +28,7 @@
 
 // hoot
 #include <hoot/core/conflate/matching/MatchFactory.h>
-#include <hoot/core/criterion/ElementTypeCriterion.h>
+#include <hoot/core/criterion/RelationCriterion.h>
 #include <hoot/core/criterion/LinearCriterion.h>
 #include <hoot/core/criterion/PointCriterion.h>
 #include <hoot/core/criterion/PolygonCriterion.h>
