+++ hoot-core/src/main/cpp/hoot/core/criterion/ElementTypeCriterion.h
@@ -62,82 +62,6 @@ protected:
   ElementType::Type _elementType;
 };
 
-class NodeCriterion : public ElementTypeCriterion
-{
-public:
-
-  static QString className() { return "hoot::NodeCriterion"; }
-
-  NodeCriterion() : ElementTypeCriterion(ElementType::Node) {}
-  ~NodeCriterion() = default;
-
-  QString getDescription() const override { return "Identifies nodes"; }
-  QString getName() const override { return className(); }
-  QString getClassName() const override { return className(); }
-  QString toString() const override { return className(); }
-};
-
-class WayCriterion : public ElementTypeCriterion
-{
-public:
-
-  static QString className() { return "hoot::WayCriterion"; }
-
-  WayCriterion() : ElementTypeCriterion(ElementType::Way) {}
-  ~WayCriterion() = default;
-
-  QString getDescription() const override { return "Identifies ways"; }
-  QString getName() const override { return className(); }
-  QString getClassName() const override { return className(); }
-  QString toString() const override { return className(); }
-};
-
-/*
- * Relations are conflatable, so RelationCriterion could inherit from ConflatableElementCriterion,
- * but it hasn't been needed so far and prevents us from multiple inheritance here.
- */
-class RelationCriterion : public ElementTypeCriterion, public Configurable
-{
-public:
-
-  static QString className() { return "hoot::RelationCriterion"; }
-
-  RelationCriterion() : ElementTypeCriterion(ElementType::Relation) {}
-  RelationCriterion(const QString& type) :
-    ElementTypeCriterion(ElementType::Relation), _type(type.trimmed()) {}
-  ~RelationCriterion() = default;
-
-  bool isSatisfied(const ConstElementPtr& e) const override
-  {
-    const bool typeMatch = ElementTypeCriterion::isSatisfied(e);
-    if (typeMatch)
-    {
-      if (_type.isEmpty())
-      {
-        return true;
-      }
-      else
-      {
-        return std::dynamic_pointer_cast<const Relation>(e)->getType() == _type;
-      }
-    }
-    return false;
-  }
-
-  QString getDescription() const override { return "Identifies relations"; }
-  QString getName() const override { return className(); }
-  QString getClassName() const override { return className(); }
-
-  void setConfiguration(const Settings& conf) override
-  {
-    _type = ConfigOptions(conf).getRelationCriterionType();
-  }
-
-private:
-
-  QString _type;
-};
-
 }
 
 #endif // ELEMENTTYPECRITERION_H
