+++ hoot-core/src/main/cpp/hoot/core/criterion/NodeCriterion.cpp
@@ -22,20 +22,20 @@
  * This will properly maintain the copyright information. Maxar
  * copyrights will be updated automatically.
  *
- * @copyright Copyright (C) 2015, 2017, 2018, 2019, 2020, 2021 Maxar (http://www.maxar.com/)
+ * @copyright Copyright (C) 2021 Maxar (http://www.maxar.com/)
  */
-#include "ElementTypeCriterion.h"
+#include "NodeCriterion.h"
 
 #include <hoot/core/util/Factory.h>
 
 namespace hoot
 {
 
-HOOT_FACTORY_REGISTER(ElementCriterion, ElementTypeCriterion)
+HOOT_FACTORY_REGISTER(ElementCriterion, NodeCriterion)
 
-bool ElementTypeCriterion::isSatisfied(const ConstElementPtr& e) const
+NodeCriterion::NodeCriterion() :
+ElementTypeCriterion(ElementType::Node)
 {
-  return e->getElementType() == _elementType;
 }
 
 }
