+++ hoot-core/src/main/cpp/hoot/core/criterion/WayCriterion.h
@@ -22,20 +22,31 @@
  * This will properly maintain the copyright information. Maxar
  * copyrights will be updated automatically.
  *
- * @copyright Copyright (C) 2015, 2017, 2018, 2019, 2020, 2021 Maxar (http://www.maxar.com/)
+ * @copyright Copyright (C) 2021 Maxar (http://www.maxar.com/)
  */
-#include "ElementTypeCriterion.h"
+#ifndef WAY_CRITERION_H
+#define WAY_CRITERION_H
 
-#include <hoot/core/util/Factory.h>
+#include <hoot/core/criterion/ElementTypeCriterion.h>
 
 namespace hoot
 {
 
-HOOT_FACTORY_REGISTER(ElementCriterion, ElementTypeCriterion)
-
-bool ElementTypeCriterion::isSatisfied(const ConstElementPtr& e) const
+class WayCriterion : public ElementTypeCriterion
 {
-  return e->getElementType() == _elementType;
-}
+public:
+
+  static QString className() { return "hoot::WayCriterion"; }
+
+  WayCriterion();
+  ~WayCriterion() = default;
+
+  QString getDescription() const override { return "Identifies ways"; }
+  QString getName() const override { return className(); }
+  QString getClassName() const override { return className(); }
+  QString toString() const override { return className(); }
+};
 
 }
+
+#endif // ELEMENTTYPECRITERION_H
