+++ hoot-core/src/main/cpp/hoot/core/visitors/MultilineStringMergeRelationCollapser.cpp
@@ -33,7 +33,7 @@
 #include <hoot/core/ops/RemoveRelationByEid.h>
 #include <hoot/core/schema/OsmSchema.h>
 #include <hoot/core/schema/MetadataTags.h>
-#include <hoot/core/criterion/ElementTypeCriterion.h>
+#include <hoot/core/criterion/RelationCriterion.h>
 #include <hoot/core/criterion/TagKeyCriterion.h>
 #include <hoot/core/elements/RelationMemberUtils.h>
